Some functional tests moved to Services sub namespace
